Message's from users

+++ ShuInkWeb/ShuInkWeb/Areas/Artist/Controllers/ArtistController.cs
@@ -1,11 +1,13 @@
-﻿namespace ShuInkWeb.Areas.Artist.Controllers
-{
-    using Microsoft.AspNetCore.Identity;
-    using Microsoft.AspNetCore.Mvc;
-    using ShuInkWeb.Core.Contracts;
-    using ShuInkWeb.Data.Entities.Identities;
-    using ShuInkWeb.Extensions;
+﻿using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using ShuInkWeb.Core.Contracts;
+using ShuInkWeb.Data.Entities.Identities;
+using ShuInkWeb.Extensions;
+using static ShuInkWeb.Constants.AreaConstants;
+using static ShuInkWeb.Constants.ActionsConstants;
 
+namespace ShuInkWeb.Areas.Artist.Controllers
+{
     public class ArtistController : BaseController
     {
         private readonly IAppointmentService appointmentService;
@@ -29,6 +31,11 @@
 
         public async Task<IActionResult> Index()
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             var id = await artistService.GetArtistIdAsync(User.Id());
 
             var models = await appointmentService.GetAppointmentsForCurrentArtist(id);
@@ -38,6 +45,11 @@
 
         public async Task<IActionResult> AllMessages()
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
             var models = await messageService.All();
 
             return View(models);
@@ -45,9 +57,19 @@
 
         public async Task<IActionResult> DeleteMessage(Guid id)
         {
+            if (!(User.IsInRole(ArtistRoleName)))
+            {
+                return RedirectToPage(InvalidOperation, new { area = IdentityRoleName });
+            }
+
+            if (!(await messageService.IsExistById(id)))
+            {
+                return RedirectToAction(nameof(AllMessages));
+            }
+
             await messageService.Delete(id);
 
-            return RedirectToAction("AllMessages", "Artist");
+            return RedirectToAction(nameof(AllMessages), ArtistControllerName);
         }
 
     }
